% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{London_Marathon_2026}
\alias{London_Marathon_2026}
\title{London Marathon 2026 Results}
\format{
A tibble with X rows and Y variables:
\describe{
  \item{Name}{Character, surname, Name}
  \item{Bib}{Numeric, Runner's Id}
  \item{Distance}{Numeric, distance in km}
  \item{Time}{Character, hh:mm:ss}
  ...
}
}
\source{
Data scraped from official results site
}
\usage{
London_Marathon_2026
}
\description{
A dataset containing the official split-level marathon results for the 2026 London Marathon of the fastest 6000 men.
}
\keyword{datasets}
