% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_race.R
\name{plot_pacing}
\alias{plot_pacing}
\title{Plot cumulative pacing for one or all runners}
\usage{
plot_pacing(data, id = NULL)
}
\arguments{
\item{data}{A tidy data frame.}

\item{id}{A Bib or Name. Use "ALL" to plot all runners.}
}
\value{
A ggplot object.
}
\description{
Plots cumulative time (hh:mm:ss) against distance. Requires columns:
Distance (numeric), Time (hh:mm:ss), and either Bib or Name.
}
